sistema_notas_pro/backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
sqlalchemy
python-multipart